fix(slab): cast XTC_SHM_HDR_SIZE to size_t in the shm-size guard (MSVC)
The v1.35.1 shm_size overflow guard compared size_t against the bare
int literal XTC_SHM_HDR_SIZE, which MSVC /W3 flags as a signed/unsigned
mismatch (C4018) and /WX makes fatal -- breaking the windows-msvc CI
build (the original addition form int + size_t had promoted to size_t,
so this was a regression introduced by the rewrite).  Cast to size_t.
Verified sign-clean with gcc -Wsign-compare -Werror across the library.

+++ src/ptc/slab.c
@@ -625,8 +625,8 @@ xtc_slab_create(const xtc_slab_opts_t *opts, xtc_slab_t **out)
 			__os_free(s);
 			return XTC_E_INVAL;
 		}
-		if (s->opts.shm_size < XTC_SHM_HDR_SIZE ||
-		    s->opts.shm_size - XTC_SHM_HDR_SIZE < s->opts.chunk_size) {
+		if (s->opts.shm_size < (size_t)XTC_SHM_HDR_SIZE ||
+		    s->opts.shm_size - (size_t)XTC_SHM_HDR_SIZE < s->opts.chunk_size) {
 			/* Region too small for header + one chunk.  Written as a
 			 * subtraction so a huge chunk_size cannot wrap the
 			 * HDR + chunk_size addition past shm_size. */
